// web/src/api/generated/types.gen.ts
export type MultiflashResponse = {
    /**
     * Phasevalues
     *
     * Phase labels (vapor, liquid, aqueous, mercury) with their fraction of unity and mercury concentration
     */
    phaseValues: {
        [key in PhaseLabels]?: {
            [key: string]: number;
        };
    };
    /**
     * Componentfractions
     *
     * Mole fractions of each of the components (Note: mass is discarded from MultiflashResult)
     */
    componentFractions: {
        [key: string]: Array<number>;
    };
    /**
     * Feedfractions
     *
     * Ratio of components in the feed (guaranteed to sum to 1)
     */
    feedFractions: {
        [key: string]: number;
    };
};

/**
 * PhaseLabels
 *
 * Enum class for phase labels
 */
export type PhaseLabels = 'Mercury' | 'Liquid' | 'Vapor' | 'Aqueous';