// web/src/api/generated/models/component-ids.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mercury
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * An enumeration.
 * @export
 * @enum {string}
 */
export enum ComponentIds {
    Id1 = 'id1',
    Id2 = 'id2',
    Id3 = 'id3',
    Id4 = 'id4',
    Id5 = 'id5',
    Id101 = 'id101',
    Id201 = 'id201',
    Id301 = 'id301',
    Id401 = 'id401',
    Id402 = 'id402',
    Id501 = 'id501',
    Id503 = 'id503',
    Id504 = 'id504',
    Id502 = 'id502',
    Id601 = 'id601',
    Id602 = 'id602',
    Id603 = 'id603',
    Id604 = 'id604',
    Id605 = 'id605',
    Id606 = 'id606',
    Id608 = 'id608',
    Id701 = 'id701',
    Id707 = 'id707',
    Id710 = 'id710',
    Id801 = 'id801',
    Id806 = 'id806',
    Id809 = 'id809',
    Id901 = 'id901',
    Id1016 = 'id1016'
}