// web/src/api/generated/models/components.ts
export interface Components {
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id1: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id2: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id3: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id4: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id5: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id101: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id201: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id301: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id401: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id402: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id501: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id503: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id504: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id502: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id601: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id602: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id603: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id604: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id605: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id606: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id608: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id701: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id707: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id710: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id801: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id806: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id809: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id901: ComponentProperties;
    /**
     * 
     * @type {ComponentProperties}
     * @memberof Components
     */
    id1016: ComponentProperties;
}