// web/src/api/generated/types/PhaseLabels.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const phaseLabelsEnum = {
    Mercury: "Mercury",
    Liquid: "Liquid",
    Vapor: "Vapor",
    Aqueous: "Aqueous"
} as const;

export type PhaseLabelsEnumKey = (typeof phaseLabelsEnum)[keyof typeof phaseLabelsEnum];

/**
 * PhaseLabels
 * @description Enum class for phase labels
*/
export type PhaseLabels = PhaseLabelsEnumKey;