// web/src/api/generated/models/phase-results.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mercury
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { PhaseProperties } from './phase-properties';

/**
 * 
 * @export
 * @interface PhaseResults
 */
export interface PhaseResults {
    /**
     * 
     * @type {PhaseProperties}
     * @memberof PhaseResults
     */
    Mercury?: PhaseProperties;
    /**
     * 
     * @type {PhaseProperties}
     * @memberof PhaseResults
     */
    Vapor?: PhaseProperties;
    /**
     * 
     * @type {PhaseProperties}
     * @memberof PhaseResults
     */
    Liquid?: PhaseProperties;
    /**
     * 
     * @type {PhaseProperties}
     * @memberof PhaseResults
     */
    Aqueous?: PhaseProperties;
}